FilterParameters class moved to models package. FilterState class moved to ViewModel.
