+++ app/src/main/java/com/academy/android/model/FilterParameters.kt
@@ -1,0 +1,9 @@
+package com.academy.android.model
+
+import com.academy.android.ui.videos.FilterState
+
+data class FilterParameters(
+    val cities: List<String>,
+    val levels: List<String>,
+    val years: List<String>
+)
