--- app/src/main/java/com/academy/android/model/FilterState.kt
@@ -1,7 +1,0 @@
-package com.academy.android.model
-
-data class FilterState(
-    val city: String,
-    val level: String,
-    val year: String,
-)
