+++ app/src/main/java/com/academy/android/model/interactors/GetFilterParametersUseCase.kt
@@ -1,7 +1,7 @@
 package com.academy.android.model.interactors
 
-import com.academy.android.data.repositories.FilterParameters
 import com.academy.android.data.repositories.VideosRepositorySource
+import com.academy.android.model.FilterParameters
 import javax.inject.Inject
 
 class GetFilterParametersUseCase @Inject constructor(
